Gaperr (#1010)
* hoc_execerr(const char* fmt, ...)
  more convenient for more complicated messages for
  hoc_execerror(const char*, const char*)

Avoid issues when creating 0 length arrays in pc.setup_transfer().

* Error message when sid < 0 for pc.source_var and pc.target_var

+++ src/oc/hoc.c
@@ -195,7 +195,7 @@ static jmp_buf hoc_oc_begin;
 int	intset;		/* safer interrupt handling */
 int	indef;
 const char	*infile;	/* input file name */
-extern int hoc_xopen_file_size_;
+extern size_t hoc_xopen_file_size_;
 extern char* hoc_xopen_file_;
 const char	**gargv;	/* global argument list */
 int	gargc;
